[#1742] Rename keys to mapping
